+++ tools/check-public-docs.sh
@@ -26,6 +26,9 @@ fi
 PATTERNS=(
   '/home/[A-Za-z0-9._-]+/'
   '/mnt/c/Users/'
+  '/mnt/data/'
+  'cyifer007@'
+  'DESKTOP-[A-Z0-9-]+'
   'x64lens_patch_[0-9]+'
   'user-created whole-repository zip snapshots'
   'in our (chat|conversation)'